Refactor dir struct. & add proper handling of actions
